+++ src/main/java/it/polimi/ingsw/is25am33/model/card/SlaveTraders.java
@@ -25,7 +25,7 @@ public class SlaveTraders extends AdvancedEnemies implements PlayerMover, CrewMe
         if (currState != GameState.CHOOSE_CANNONS)
             throw new IllegalStateException("Not the right state");
 
-        int currPlayerCannonPower = activateDoubleCannonsProcess(chosenDoubleCannons, chosenBatteryBoxes, game.getCurrPlayer());
+        double currPlayerCannonPower = activateDoubleCannonsProcess(chosenDoubleCannons, chosenBatteryBoxes, game.getCurrPlayer());
 
         if (currPlayerCannonPower > requiredFirePower) {
